requirements.txt
streamlit
requests
beautifulsoup4
reportlab
pandas
pillow
